Fixed bug in StringToken that returned the index as a double.

+++ src/parser/StringToken.java
@@ -8,7 +8,7 @@ public class StringToken extends Token{
 		index = val;
 	}
 	
-	public double getIndex() {
+	public int getIndex() {
 		return index;
 	}
 }
